python/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
tapsilat-py==2025.6.24.1
gunicorn==21.2.0
